#281 implement optimum-checkbox to set pdfs per-page to 0
New std value is 0 (but current value will be saved as user preference)
and print-previews per-page setting is now uncorrelated with submissions-
tabs per-page setting.
Furthermore the checkbox labeled optimum overrides the value in the
textfield to set it automatically to 0.

JS which disables the textfield as long as optimum-checkbox is checked
is not functional at the moment.