#5426 fix non groups members staying selected for export if selected group is changed before pressing export
If someone selects users in the preview table to be exported and then changes displayed group the old users stay selected even if only the newly selected group should be exported which causes an error. We now remove all users from the selection if they're not part of the currently selected group.